+++ docs/theme/yumemi.js
@@ -22,6 +22,7 @@
             { id: "guzzle", title: "Guzzle" },
             { id: "getid3", title: "getID3" },
             { id: "illuminate-auth", title: "Illuminate Auth" },
+            { id: "illuminate-bus", title: "Illuminate Bus" },
             { id: "illuminate-cache", title: "Illuminate Cache" },
             { id: "illuminate-console", title: "Illuminate Console" },
             { id: "illuminate-cookie", title: "Illuminate Cookie" },