+++ nix/consumer-cache-hash.nix
@@ -1,1 +1,1 @@
-"sha256-1o7sNNt351B0R+cWveyQit7vTIO96a4UW7yvQX2gU+s="
+"sha256-VsYoVgIiJac+AbGK6SjuPTar/zibVBUyqDkP0yLOXwQ="